# documentation/openapi/providers.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "providers API",
    "description": "Internal HTTP API for the providers service.",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/admin/providers/{provider}/capability-refresh": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "provider",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "settings"
                  ],
                  "properties": {
                    "settings": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "provider",
                          "capability",
                          "refreshIntervalMs",
                          "saveResolutionMs",
                          "enabled"
                        ],
                        "properties": {
                          "provider": {
                            "type": "string"
                          },
                          "capability": {
                            "type": "string"
                          },
                          "refreshIntervalMs": {
                            "type": "integer"
                          },
                          "saveResolutionMs": {
                            "anyOf": [
                              {
                                "type": "integer"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "enabled": {
                            "type": "boolean"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/providers/{provider}/capability-refresh/{capability}": {
      "put": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refresh_interval_ms": {
                    "minimum": 1000,
                    "maximum": 2592000000,
                    "type": "integer"
                  },
                  "save_resolution_ms": {
                    "anyOf": [
                      {
                        "minimum": 1000,
                        "maximum": 2592000000,
                        "type": "integer"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "enabled": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "provider",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "capability",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "provider",
                    "capability",
                    "refreshIntervalMs",
                    "saveResolutionMs",
                    "enabled"
                  ],
                  "properties": {
                    "provider": {
                      "type": "string"
                    },
                    "capability": {
                      "type": "string"
                    },
                    "refreshIntervalMs": {
                      "type": "integer"
                    },
                    "saveResolutionMs": {
                      "anyOf": [
                        {
                          "type": "integer"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "enabled": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/providers/{provider}/search": {
      "get": {
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "maxLength": 120,
              "type": "string"
            },
            "in": "query",
            "name": "q",
            "required": true
          },
          {
            "schema": {
              "minimum": 1,
              "maximum": 25,
              "type": "integer"
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "provider",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "provider",
                    "results"
                  ],
                  "properties": {
                    "provider": {
                      "type": "string"
                    },
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "symbol",
                          "name",
                          "exchange",
                          "mic",
                          "currency",
                          "quoteType"
                        ],
                        "properties": {
                          "symbol": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "exchange": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "mic": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "currency": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "quoteType": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/providers/{provider}": {
      "patch": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "enabled": {
                    "type": "boolean"
                  },
                  "data_quality": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "high"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "medium"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "low"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "unknown"
                        ]
                      }
                    ]
                  },
                  "capability_quality": {
                    "type": "object",
                    "additionalProperties": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "high"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "medium"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "low"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "unknown"
                          ]
                        }
                      ]
                    }
                  },
                  "max_batch_size": {
                    "anyOf": [
                      {
                        "minimum": 1,
                        "type": "integer"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "rate_limit_per_min": {
                    "anyOf": [
                      {
                        "minimum": 1,
                        "type": "integer"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "max_concurrency": {
                    "minimum": 1,
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "provider",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "provider",
                    "enabled",
                    "providerClass",
                    "dataQuality",
                    "capabilityQuality",
                    "maxBatchSize",
                    "rateLimitPerMin",
                    "maxConcurrency"
                  ],
                  "properties": {
                    "provider": {
                      "type": "string"
                    },
                    "enabled": {
                      "type": "boolean"
                    },
                    "providerClass": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "symbol"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "reference"
                          ]
                        }
                      ]
                    },
                    "dataQuality": {
                      "anyOf": [
                        {
                          "type": "string",
                          "enum": [
                            "high"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "medium"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "low"
                          ]
                        },
                        {
                          "type": "string",
                          "enum": [
                            "unknown"
                          ]
                        }
                      ]
                    },
                    "capabilityQuality": {
                      "type": "object",
                      "additionalProperties": {
                        "anyOf": [
                          {
                            "type": "string",
                            "enum": [
                              "high"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "medium"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "low"
                            ]
                          },
                          {
                            "type": "string",
                            "enum": [
                              "unknown"
                            ]
                          }
                        ]
                      }
                    },
                    "maxBatchSize": {
                      "anyOf": [
                        {
                          "type": "integer"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "rateLimitPerMin": {
                      "anyOf": [
                        {
                          "type": "integer"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "maxConcurrency": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/internal/analyst": {
      "get": {
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "maxLength": 40,
              "type": "string"
            },
            "in": "query",
            "name": "symbol",
            "required": true
          },
          {
            "schema": {
              "minLength": 1,
              "maxLength": 64,
              "type": "string"
            },
            "in": "query",
            "name": "provider",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "provider",
                    "assessment"
                  ],
                  "properties": {
                    "provider": {
                      "type": "string"
                    },
                    "assessment": {}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/internal/fundamentals": {
      "get": {
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "maxLength": 40,
              "type": "string"
            },
            "in": "query",
            "name": "symbol",
            "required": true
          },
          {
            "schema": {
              "minLength": 1,
              "maxLength": 64,
              "type": "string"
            },
            "in": "query",
            "name": "provider",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "provider",
                    "fundamentals"
                  ],
                  "properties": {
                    "provider": {
                      "type": "string"
                    },
                    "fundamentals": {}
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/internal/news": {
      "get": {
        "parameters": [
          {
            "schema": {
              "minLength": 1,
              "maxLength": 40,
              "type": "string"
            },
            "in": "query",
            "name": "symbol",
            "required": true
          },
          {
            "schema": {
              "minLength": 1,
              "maxLength": 64,
              "type": "string"
            },
            "in": "query",
            "name": "provider",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "provider",
                    "news"
                  ],
                  "properties": {
                    "provider": {
                      "type": "string"
                    },
                    "news": {}
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}